Moved message, message pool, proposed block to consensus.
